requirements.txt
PySide6>=6.7
# Core dependencies
numpy>=1.24.0
laspy>=2.5.0
open3d>=0.17.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# GUI dependencies
tkinter  # Usually comes with Python
sv-ttk>=1.0.0

# Visualization and analysis
matplotlib>=3.7.0
alphashape>=1.3.0
google-generativeai>=0.3.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
